reorganize imports around u in top pkg
