+++ particula/aerosol_dynamics/__init__.py
@@ -1,11 +1,13 @@
-"""module to perform calculations related to aerosol dynamics.
+""" module to perform calculations related to aerosol dynamics.
 
-This module contains functions to perform calculations related to
-aerosol dynamics. It inherits the unit registry, u, from particula.
-The pacakge contents below are callablle. For example, the particle
-module is callable as:
+    This module contains functions to perform calculations related
+    to aerosol dynamics. It inherits the unit registry, u, fro,
+    particula. The pacakge contents below are callablle. For example,
+    the particle module is callable as:
 
-    from particula.aerosol_dynamics import particle
+        from particula.aerosol_dynamics import particle
 
-More details to follow.
+    ...
+
+    More details to follow.
 """
