+++ particula/aerosol_dynamics/parcel.py
@@ -1,8 +1,10 @@
 """
 Class for creating a air parcel to put particles and gases in.
 """
+
 import numpy as np
-from particula.aerosol_dynamics import particle, u
+from particula import u
+from particula.aerosol_dynamics import particle
 
 
 class Parcel:
