+++ particula/aerosol_dynamics/tests/particle_test.py
@@ -5,10 +5,8 @@ Test suites for particle class.
 import numpy as np
 import pint
 import pytest
-
+from particula import u
 from particula.aerosol_dynamics import environment, particle
-
-from particula.aerosol_dynamics import u
 
 small_particle = particle.Particle(
     name="small_particle",
